Push switch element into separate file
