+++ client/src/style/scss/ui.scss
@@ -288,8 +288,8 @@ $ui-margin-horizontal-large: $margin-v * 2;
         @extend .b-custom-control-lg;
         @extend .float-left;
         @extend .no-highlight;
-        input {
-            @extend .custom-control-input;
+        .custom-control-label:before {
+            border-color: $border-color;
         }
     }
     .label {
